Make Amount<C, T> polymorphic over the backing representation T.
Because we may wish to represent currency amounts in a variety of ways,
we now make Amount<C, T> take a T parameter which indicates the
underlying numeric type used for computations. It's expected that this
is something into which we can inject u64, and from which we can (maybe)
extract a u64.